feat(trips): move trip actions into the header
