+++ frontend/features/trips/presentation/trip-layout-client.tsx
@@ -46,7 +46,7 @@ function TripShell({ children }: { children: React.ReactNode }) {
     <>
       {portalTarget && createPortal(<TripTabBar />, portalTarget)}
       <div>
-        <TripHeader trip={data.trip} />
+        <TripHeader />
         <div className="border-t border-border/40 px-4 pb-4 pt-4 sm:px-6 sm:pb-6">
             {children}
           </div>
